rule_based_system/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
pydantic>=2.0.0
spacy>=3.7.0
pandas>=2.0.0
openpyxl>=3.1.0
pdfplumber>=0.11.0
python-docx>=1.1.0
